+++ Makefile
@@ -10,6 +10,7 @@ all: maru2 maru2-publish ## Build all binaries
 maru2: ## Build maru2 binary and generate schemas
 	go build -o bin/ -ldflags="-s -w" -trimpath ./cmd/maru2
 	go run cmd/maru2-schema/main.go v0 > schema/v0/schema.json
+	go run cmd/maru2-schema/main.go v1 > schema/v1/schema.json
 	go run cmd/maru2-schema/main.go > maru2.schema.json
 
 maru2-publish: ## Build maru2-publish binary
