+++ schema/v1/doc.go
@@ -1,7 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2025-Present Defense Unicorns
 
-// Package v0 is the v0 (alpha) schema for maru2,
-//
-// Breaking changes are to be expected without any migration pathway.
-package v0
+// Package v1 is the v1 schema for maru2 workflows
+package v1
