+++ schema/v1/uses.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2025-Present Defense Unicorns
 
-package v0
+package v1
 
 // SupportedSchemes returns a list of supported schemes
 func SupportedSchemes() []string {
